+++ setupmeta/commands.py
@@ -16,12 +16,6 @@ import setupmeta
 flatten = chain.from_iterable
 
 
-def abort(message):
-    from distutils.errors import DistutilsSetupError
-
-    raise DistutilsSetupError(message)
-
-
 def MetaCommand(cls):
     """Decorator allowing for less boilerplate in our commands"""
     return setupmeta.MetaDefs.register_command(cls)
@@ -43,21 +37,19 @@ class CheckCommand(check_cmd):
 
     user_options = check_cmd.user_options + [
         ("status", "t", "Show git status recap (useful to get evidence as to why version was dirty during CI jobs)"),
-        ("deptree", "d", "Show dependency tree (from currently activated venv, or ./.venv, or ./venv)"),
         ("reqs", "q", "Show how many requirements were auto-abstracted or ignored, if any"),
     ]
 
     def initialize_options(self):
         check_cmd.initialize_options(self)
         self.status = None
-        self.deptree = None
         self.reqs = None
 
     def run(self):
         if not self.setupmeta:
             return check_cmd.run(self)
 
-        if count(self.restructuredtext, self.status, self.deptree, self.reqs) == 0:
+        if count(self.restructuredtext, self.status, self.reqs) == 0:
             self.status = 1
             self.reqs = 1
 
@@ -66,9 +58,6 @@ class CheckCommand(check_cmd):
 
         if self.status:
             self._show_git_status()
-
-        if self.deptree:
-            self._warnings = _show_dependencies(self.setupmeta.definitions)
 
         check_cmd.run(self)
 
@@ -126,7 +115,9 @@ class VersionCommand(setuptools.Command):
                 print(self.setupmeta.version)
 
         except setupmeta.UsageError as e:
-            abort(e)
+            from distutils.errors import DistutilsSetupError
+
+            raise DistutilsSetupError(e)
 
 
 @MetaCommand
@@ -428,244 +419,3 @@ class CleanCommand(setuptools.Command):
 
         if self.deleted == 0:
             print("all clean, no deletable files found")
-
-
-def _show_dependencies(definitions):
-    """
-    Conveniently  get dependency tree via ./setup.py check --dep, similar to https://pypi.org/project/pipdeptree
-    """
-    if setupmeta.pkg_resources is None or not hasattr(setupmeta.pkg_resources, "WorkingSet"):
-        setupmeta.warn("pkg_resources is not available, can't show dependencies")
-        return 1
-
-    venv = find_venv()
-    if not venv:
-        setupmeta.warn("Could not find virtual environment to scan for dependencies")
-        return 1
-
-    entries = list(find_subfolders(venv, ["site-packages"]))
-    if not entries:
-        setupmeta.warn("Could not find 'site-packages' subfolder in '%s'" % venv)
-        return 1
-
-    tree = DepTree(setupmeta.pkg_resources.WorkingSet(entries), definitions)
-    print(tree.rendered())
-    return len(tree.conflicts) + len(tree.cycles)
-
-
-class PipReq(object):
-    def __init__(self, obj, package):
-        """
-        :param pkg_resources.Requirement obj:
-        :param PipPackage package: Associated package
-        """
-        self._obj = obj
-        self.package = package
-        self.key = obj.key
-        self.version = package.version
-        self.version_spec = ",".join(["".join(sp) for sp in sorted(obj.specs, reverse=True)])
-        self.version_rec = setupmeta.pkg_req("%s%s" % (self.key, self.version_spec or ""))
-        self.is_conflicting = not self.package.version or self.package.version not in self.version_rec
-
-    def __repr__(self):
-        return self.key
-
-    def __eq__(self, other):
-        return isinstance(other, PipReq) and self.key is other.key
-
-    def __lt__(self, other):
-        return isinstance(other, PipReq) and self.key < other.key
-
-    def render(self):
-        conflict = " CONFLICT!" if self.is_conflicting else ""
-        return "%s [required: %s, installed: %s]%s" % (self.key, self.version_spec or "Any", self.version, conflict)
-
-
-class PipPackage(object):
-    """Represents a pip package"""
-
-    def __init__(self, tree, obj):
-        """
-        :param DepTree tree: Associated tree
-        :param pkg_resources.DistInfoDistribution obj:
-        """
-        self.tree = tree
-        self._obj = obj
-        self.key = obj.key
-        self.version = obj.version
-        self.requires = []
-        self.required_by = set()
-        self.transitive = set()
-        self.cycle = None
-
-    def __repr__(self):
-        return self.key
-
-    def __eq__(self, other):
-        return isinstance(other, PipPackage) and self.key is other.key
-
-    def __lt__(self, other):
-        return isinstance(other, PipPackage) and self.key < other.key
-
-    def __hash__(self):
-        return hash(self.key)
-
-    def resolve(self):
-        for req in self._obj.requires():
-            package = self.tree.get_package(req)
-            if package:
-                pr = PipReq(req, package)
-                self.requires.append(pr)
-                package.required_by.add(self)
-
-    def _add_transitive(self, required):
-        if isinstance(required, PipReq):
-            required = required.package
-
-        if isinstance(required, PipPackage):
-            if required not in self.transitive:
-                self.transitive.add(required)
-                self._add_transitive(required.requires)
-
-            return
-
-        for req in required:
-            self._add_transitive(req)
-
-    def _find_cycle(self, target, visited):
-        if self in visited:
-            return None
-
-        visited.add(self)
-        for r in sorted(self.requires):
-            if r.package is target:
-                return [r.package]
-
-            c = r.package._find_cycle(target, visited)
-            if c:
-                return [r.package] + c
-
-    def resolve_transitive(self):
-        self._add_transitive(self.requires)
-        if self in self.transitive:
-            self.cycle = self._find_cycle(self, set())
-
-    def render(self):
-        return "%s==%s" % (self.key, self.version)
-
-
-def find_subfolders(folder, names, depth=3):
-    if folder and os.path.isdir(folder):
-        for name in os.listdir(folder):
-            fpath = os.path.join(folder, name)
-            if name in names:
-                yield fpath
-                continue
-
-            if os.path.isdir(fpath) and depth > 0:
-                for p in find_subfolders(fpath, names, depth=depth - 1):
-                    yield p
-
-
-def find_venv():
-    venv = os.environ.get("VIRTUAL_ENV")
-    if venv:
-        return venv
-
-    for folder in (".venv", "venv"):
-        fpath = setupmeta.project_path(folder)
-        if os.path.isdir(fpath):
-            return fpath
-
-
-class DepTree:
-    def __init__(self, ws, definitions):
-        self.packages = dict((d.key, PipPackage(self, d)) for d in ws)
-        self.setup = definitions.get("setup_requires")
-        self.install_requires = definitions.get("install_requires")
-        self.extras_require = definitions.get("extras_require")
-        self.conflicts = set()
-        self.cycles = {}
-
-        for p in sorted(self.packages.values()):
-            p.resolve()
-            self.conflicts.update(r.key for r in p.requires if r.is_conflicting)
-
-        for p in sorted(self.packages.values()):
-            p.resolve_transitive()
-            if p.cycle:
-                key = setupmeta.represented_args(sorted(p.cycle))
-                if key not in self.cycles:
-                    self.cycles[key] = [p] + p.cycle
-
-    def get_package(self, ref):
-        return self.packages.get(getattr(ref, "key", ref))
-
-    def get_packages(self, dependencies):
-        result = []
-        for dep in dependencies:
-            p = self.get_package(setupmeta.pkg_req(dep))
-            if p:
-                result.append(p)
-
-        return result
-
-    def get_children(self, ref):
-        return self.get_package(ref).requires
-
-    def render_section(self, report, seen, title, dependencies):
-        nodes = self.get_packages(dependencies)
-        if not nodes:
-            return
-
-        def aux(node, indent=2, chain=None):
-            if chain is None:
-                chain = []
-
-            result = ["%s%s" % (" " * indent, node.render())]
-            children = sorted(self.get_children(node))
-            children = [aux(c, indent=indent + 2, chain=chain + [c.key]) for c in children if c.key not in chain]
-
-            chain.append(node.key)
-            p = self.packages.get(node.key)
-            if p:
-                seen.add(p)
-
-            result += list(flatten(children))
-            return result
-
-        seen.update(nodes)
-        auxed = [aux(p) for p in nodes]
-        report.append("%s:\n%s" % (title, "-" * len(title)))
-        report.extend(flatten(auxed))
-        report.append("")
-
-    def rendered(self):
-        """String representation"""
-        result = ["Dependency tree:"]
-        seen = set()
-
-        if self.install_requires:
-            self.render_section(result, seen, "install_requires", self.install_requires.value)
-
-        if self.extras_require and self.extras_require.value:
-            for name, value in self.extras_require.value.items():
-                self.render_section(result, seen, "extras_require[%s]" % name, value)
-
-        other = set(self.packages.values()) - seen
-        if other:
-            other = sorted(p.key for p in other if not p.required_by)
-            self.render_section(result, seen, "other", other)
-
-        if self.conflicts:
-            result.append("\n%s conflicts: %s" % (len(self.conflicts), setupmeta.represented_args(self.conflicts, separator=", ")))
-
-        if self.cycles:
-            result.append("\n%s cycles found:" % len(self.cycles))
-            for c in sorted(self.cycles.values()):
-                result.append(setupmeta.represented_args(c, separator=" -> "))
-
-        if len(result) < 2:
-            result.append("- no dependencies -")
-
-        return "\n".join(result)